requirements.txt
aiohttp==3.8.1
aiosignal==1.2.0
asgiref==3.5.0
async-timeout==4.0.2
asynctest==0.13.0
attrs==21.4.0
beautifulsoup4==4.10.0
bs4==0.0.1
certifi==2021.10.8
charset-normalizer==2.0.12
click==8.0.4
colorama==0.4.4
distlib==0.3.4
Django==3.2.12
filelock==3.6.0
Flask==2.0.3
frozenlist==1.3.0
future==0.18.2
gunicorn==20.1.0
idna==3.3
importlib-metadata==4.11.3
itsdangerous==2.1.1
Jinja2==3.0.3
line-bot-sdk==2.1.0
MarkupSafe==2.1.1
multidict==6.0.2
numpy==1.21.6
pandas==1.3.5
platformdirs==2.5.1
python-dateutil==2.8.2
pytz==2022.1
requests==2.27.1
six==1.16.0
soupsieve==2.3.1
sqlparse==0.4.2
typing-extensions==4.1.1
urllib3==1.26.9
virtualenv==20.14.0
Werkzeug==2.0.3
yarl==1.7.2
zipp==3.7.0
